api/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.109.*
uvicorn[standard]==0.27.*
pydantic==2.*
google-cloud-storage==2.*
google-auth==2.*
requests==2.*
python-multipart==0.0.*
